fixup CertificateType Enum to follow RFC naming standards
Also changed the RcordToString() of CertRecord from outputting unsigned integer to the mnemonic representation for both algorithm and certtype.  This is allowed according to [RFC 4398 Section 2.2](https://datatracker.ietf.org/doc/html/rfc4398#section-2.2)

+++ src/DnsClient/Protocol/CertRecord.cs
@@ -79,6 +79,6 @@ public class CertRecord : DnsResourceRecord
     
     private protected override string RecordToString()
     {
-        return string.Format("{0} {1} {2} {3}", (int)CertType, KeyTag, (int)Algorithm, PublicKeyAsString);
+        return string.Format("{0} {1} {2} {3}", CertType, KeyTag, Algorithm, PublicKeyAsString);
     }
 }
